week_2_data_ingestion/homework/requirements.txt
apache-airflow-providers-google
apache-airflow-providers-amazon
boto3
pyarrow
pydantic
